feat: add map label printing
